requirements/base.txt
django
pillow
djangorestframework
djangorestframework-simplejwt
python-decouple
django-redis
celery
psycopg2-binary
drf-spectacular
django-import-export
colorlog
djangorestframework
django-axes
django-ratelimit
drf-spectacular
drf-nested-routers
celery[redis]
django-celery-beat
django-celery-results
flower
firebase-admin
pyfcm